+++ README.md
@@ -1,28 +1,39 @@
 # Телефонный справочник
 
-## Главное меню
-        --------------+-------------+-----------+------------------+------------+-----------+------------+----------
-        | 1. Просмотр | 2. Добавить | 3. Искать | 4. Редактировать | 5. Удалить | 6. Импорт | 7. Экспорт | 8. Exit |
-        --------------+-------------+-----------+------------------+------------+-----------+------------+----------
+## Описание задачи.
+Создать телефонный справочник с возможностью импорта и экспорта данных в нескольких форматах.
+Под форматами понимаем структуру файлов, например: в файле на одной строке хранится одна часть записи, пустая строка - разделитель.
 
-1. Просмотр. Данный модуль выводит в консоль список всех имеющихся в базе контактов 
+---
+## Команда
 
-2. Добавить. Реализует возможность добавления новых конактов
+- Женя (https://github.com/pogudo-e)
+- Карэн (https://github.com/karenyaylyan)
 
-3. Искать. Поиск контакта по базе контактов
+---
+## Структура приложения (модули)
 
-4. Редактировать. В процессе.
+- main - Точка входа в программу (*Женя*)
+- logger - Логироваине данных: add, del, edit (*Карэн*)
+- html_creater - Генерирует html файл, сохраняет в папке export (*Женя*)
+- json_creater - Генерирует json файл сохраняет в папке export (*Карэн*)
+- import_contacts - Испортирует данные из указанного файла с разделителем (перевод строки), Присваивает уникальный id и построчно добавляет в БД (*Женя*)
+- database.txt - База данных содержащая в себе уникальный  id, name, phone, email. Разделителем служит перевод строки (*Карэн*)
+- add_contact - Добавление нового контакта (*Женя*)
+- edit_contact - Редактироваине существующего контакта по id (*Карэн*)
+- del_contact - Удаление контакта по id (*Карэн*)
+- view_contacts - Вывод полного списка в консоль (*Женя*)
 
-5. Удалить. Возможность удалить контакт по уникальному идентификатору
+---
+## Как запустить проект (рекомендуем использовать VSC)
 
-6. Импорт. Возможность добавить в базу контактов другие контакты указав пусть к файлу
+### Для запуска с помощью консоли:
+- Скачать проект в локальный репозиторий с помощью команды:
 
-7. Экспорт:
+        git clone https://github.com/karenyaylyan/gb_notebook.git
 
-    7.1.  Exel. Обсуждается т.к. сложная реализация без сторонних модулей
+- В консоли ввести команду 
+        
+        python3 main.py
 
-    7.2. Html
-
-    7.3. Json
-
-8. Exit. Выход из программы
+- Следовать инструкциям в консоли.
