+++ console/view.py
