--- export/09-12-2022-01-00.html
@@ -1,63 +1,0 @@
-<html>
-  <head><style>table{font-family:arial,sans-serif;border-collapse:collapse;width:100%}td,th{border:1px solid #ddd;text-align:left;padding:8px}tr:nth-child(even){background-color:#ddd}</style></head>
-  <body>
-<h2>HTML Table</h2><table>
-<tr><th>ID</th><th>Name</th><th>Phone</th><th>Email</th></tr><tr>
-<td>1</td>
-<td>Андрей</td>
-<td>+7(999)123-45-67</td>
-<td>an@mail.ru</td>
-</tr><tr>
-<td>2</td>
-<td>Екатерина</td>
-<td>+7(999)223-45-67</td>
-<td>ek@mail.ru</td>
-</tr><tr>
-<td>3</td>
-<td>Анастасия</td>
-<td>+7(999)323-45-67</td>
-<td>ana@mail.ru</td>
-</tr><tr>
-<td>4</td>
-<td>Влад</td>
-<td>+7(999)423-45-67</td>
-<td>vl@gmail.ru</td>
-</tr><tr>
-<td>5</td>
-<td>Валерий</td>
-<td>+7(999)523-45-67</td>
-<td>va@gmail.ru</td>
-</tr><tr>
-<td>6</td>
-<td>Михаил</td>
-<td>+7(999)623-45-67</td>
-<td>mi@gmail.ru</td>
-</tr><tr>
-<td>7</td>
-<td>Степан</td>
-<td>+7(999)723-45-67</td>
-<td>st@gmail.ru</td>
-</tr><tr>
-<td>9</td>
-<td>Мария</td>
-<td>+7(999)923-45-67</td>
-<td>mar@gmail.ru</td>
-</tr><tr>
-<td>10</td>
-<td>Анастасия</td>
-<td>+7(999)224-45-67</td>
-<td>anasteisha@gmail.ru</td>
-</tr><tr>
-<td>11</td>
-<td>Женя</td>
-<td>+7(999)224-45-67</td>
-<td>evg@mail.ru</td>
-</tr><tr>
-<td>12</td>
-<td>Михаил</td>
-<td>+7(999)623-45-67</td>
-<td>mi@gmail.ru</td>
-</tr>
-</table>
-</body>
-</html>
