--- export/09-12-2022-01-18.json
@@ -1,10 +1,0 @@
-{"id": "2", "name": "Екатерина", "phone": "+7(999)223-45-67", "mail": "ek@mail.ru"}
-{"id": "3", "name": "Анастасия", "phone": "+7(999)323-45-67", "mail": "ana@mail.ru"}
-{"id": "4", "name": "Влад", "phone": "+7(999)423-45-67", "mail": "vl@gmail.ru"}
-{"id": "5", "name": "Валерий", "phone": "+7(999)523-45-67", "mail": "va@gmail.ru"}
-{"id": "6", "name": "Михаил", "phone": "+7(999)623-45-67", "mail": "mi@gmail.ru"}
-{"id": "7", "name": "Степан", "phone": "+7(999)723-45-67", "mail": "st@gmail.ru"}
-{"id": "9", "name": "Мария", "phone": "+7(999)923-45-67", "mail": "mar@gmail.ru"}
-{"id": "10", "name": "Анастасия", "phone": "+7(999)224-45-67", "mail": "anasteisha@gmail.ru"}
-{"id": "11", "name": "Женя", "phone": "+7(999)224-45-67", "mail": "evg@mail.ru"}
-{"id": "12", "name": "Михаил", "phone": "+7(999)623-45-67", "mail": "mi@gmail.ru"}
