+++ log.txt
@@ -26,3 +26,6 @@
 
 02:02 edit 1 Светлана +7(909)543-23-23 sv@mail.ru
 
+10:02 del 12 Михаил +7(999)623-45-67 mi@gmail.ru
+10:02 add 12 Михаил +7(999)623-45-67 mi@gmail.ru
+
